feat: refactor variant navigation logic and enhance RepertoireInfoPanel with training action
