+++ packages/frontend/src/components/application/chess/board/RepertoireInfo.tsx
@@ -1,14 +1,11 @@
 import React from "react";
-import { useNavigate, useLocation } from "react-router-dom";
 import { RepertoireInfoPanel } from "../../../design/chess/RepertoireInfoPanel/RepertoireInfoPanel";
 import { useRepertoireContext } from "../../../../contexts/RepertoireContext";
 import { useRepertoireInfo } from "../../../../hooks/useRepertoireInfo";
 import { useMenuContext } from "../../../../contexts/MenuContext";
-import { Variant } from "../../../../models/chess.models";
+import { useVariantNavigation } from "../../../../hooks/useVariantNavigation";
 
 export const RepertoireInfo = () => {
-  const navigate = useNavigate();
-  const location = useLocation();
   const {
     variants,
     currentMoveNode,
@@ -19,27 +16,11 @@ export const RepertoireInfo = () => {
     comment,
     updateComment,
     selectedVariant,
-    setSelectedVariant,
     repertoireId,
-    initBoard,
   } = useRepertoireContext();
 
   const { toggleMenu } = useMenuContext();
-
-  const handleVariantChange = (variant: Variant | null) => {
-    setSelectedVariant(variant);
-    initBoard();
-    
-    const currentParams = new URLSearchParams(location.search);
-    if (variant?.fullName) {
-      currentParams.set("variantName", variant.fullName);
-    } else {
-      currentParams.delete("variantName");
-    }
-    
-    const newUrl = `${location.pathname}?${currentParams.toString()}`;
-    navigate(newUrl, { replace: true });
-  };
+  const { handleVariantChange } = useVariantNavigation();
 
   const { downloadVariantPGN, copyVariantPGN,copyVariantToRepertoire,copyVariantsToRepertoire,deleteVariants, deleteVariant } =
     useRepertoireInfo();
